requirements.txt
python-telegram-bot==20.3
yt-dlp
python-dotenv==1.0.1
